DEV: partially bake in env
It may be useful for launcher2 build to bake in some env, allowing for default
env to run from a built image.

Introduce defaultBakeEnv as a list of env to bake into images. This allows for
folks pulling images to not need to mark every env variable, allowing for some
convenience in starting images.