Add multi-stage builder
Base: add pnpm deps that stores the pnpm store.
Remove the pnpm store from discourse_slim.

web.template:
Add the ability to run a multi-stage builder, so we can do a pnpm and compile
step separately. Allow the ability to copy the result into a different place.

Add a step that copies the output of a compiled app into
/var/www/app/assets/javascripts/discourse/dist from a pre-compiled build.